correct datatype for img

+++ hyperspec/transforms/image_transform.py
@@ -94,8 +94,8 @@ class ImageTransform(BaseTransform):
         zeroPaddedX = self._pad_with_zeros(X, margin=margin)  # 3-dim numpy array of size (x+margin, y+margin, z)
         total_split = X.shape[0] * X.shape[1]
 
-        patchesData = np.zeros((X.shape[0] * X.shape[1], self.windowSize, self.windowSize, X.shape[2]), dtype=np.int16)
-        patchesLabels = np.zeros((X.shape[0] * X.shape[1]), dtype=np.int16)
+        patchesData = np.zeros((X.shape[0] * X.shape[1], self.windowSize, self.windowSize, X.shape[2]), dtype=np.float64)
+        patchesLabels = np.zeros((X.shape[0] * X.shape[1]), dtype=np.float64)
         patchIndex = 0
 
         for r in range(margin, zeroPaddedX.shape[0] - margin):
